Created gazebo_utils and math_utils APIs, added comments
